requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
faster-whisper>=1.0.0
piper-tts>=1.2.0
requests>=2.31.0
pydantic>=2.5.0
numpy>=1.26.0
PyYAML>=6.0.1
python-multipart>=0.0.9
